make changes in source code files,endure implementation of func.

+++ src/libgeometry/CircleClass.cpp
@@ -8,7 +8,5 @@ void CircleClass::SetXcoord(float x) { xCoord = x; }
 void CircleClass::SetYcoord(float y) { yCoord = y; }
 void CircleClass::SetRadious(float rad) { radious = rad; }
 
-double CircleClass::CalculatePerimeter(float r) { return  2 * r * M_PI; }
-double CircleClass::CalculateArea(float r) { return r * r * M_PI; }
-
-
+float CircleClass::CalculatePerimeter(float r) { return  2 * r * M_PI; }
+float CircleClass::CalculateArea(float r) { return r * r * M_PI; }
